jugadores y publicaciones funcional

+++ backend/app/Http/Controllers/Api/JugadorController.php
@@ -18,7 +18,14 @@ class JugadorController extends Controller
      */
     public function index()
     {
-        return JugadorResource::collection(Jugador::all());
+        $jugadores = Jugador::with([
+            'imagenes',
+            'equipo',
+            'actas',
+            'publicaciones',
+            'estudio'
+        ])->get();
+        return JugadorResource::collection($jugadores);
     }
 
     /**
@@ -34,22 +41,41 @@ class JugadorController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Jugador $jugador)
+    public function show($id)
     {
-        return new JugadorResource($jugador);
+        $jugadores = Jugador::with([
+            'imagenes',
+            'equipo',
+            'actas',
+            'estudio',
+            'publicaciones',
+
+        ])->find($id);
+
+        if (!$jugadores) {
+            return response()->json(['error' => 'Jugador no encontrado'], 404);
+        }
+
+        return new JugadorResource($jugadores);
     }
 
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateJugadorRequest $request, Jugador $jugador)
+    public function update(UpdateJugadorRequest $request, $id)
     {
+        // Buscar el jugador por el ID
+        $jugador = Jugador::find($id);
+        // Verificar si el jugador existe
+        if (!$jugador) {
+            return response()->json(['error' => 'Jugador no encontrado'], 404);
+        }
+        // Validar los datos del request
         $datos = $request->validated();
-        Log::info('Datos validados:', $datos);
-
+        // Actualizar los datos del jugador
         $jugador->update($datos);
-
+        // Devolver una respuesta con el jugador actualizado
         return response()->json(['message' => 'Jugador actualizado correctamente', 'data' => $jugador]);
     }
 
@@ -57,8 +83,12 @@ class JugadorController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Jugador $jugador)
+    public function destroy($id)
     {
+        $jugador = Jugador::find($id);
+        if (!$jugador) {
+            return response()->json(['error' => 'Jugador no encontrado'], 404);
+        }
         $jugador->delete();
         return response()->noContent();
     }
