+++ backend/app/Http/Controllers/Api/PublicacionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Requests\PublicacionRequests\StorePublicacionRequest;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\PublicacionRequests\UpdatePublicacionRequest;
 use App\Http\Resources\PublicacionResource;
 use App\Models\Publicacion;
 use Illuminate\Http\Request;
@@ -16,49 +17,79 @@ class PublicacionController extends Controller
      */
     public function index()
     {
-        $publicaciones = Publicacion::with('imagenes,equipo,partido,patrocinador,jugador,reto,ong,pabellon')->get();
-        return PublicacionResource::collection($publicaciones);
+            $publicaciones = Publicacion::with([
+                'imagenes',
+                'equipo',
+                'partido',
+                'patrocinador',
+                'jugador',
+                'reto',
+                'ong',
+                'pabellon'
+            ])->get();
+            return PublicacionResource::collection($publicaciones);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(StorePublicacionRequest $request)
     {
-        dd($request);
-        $publicacion= new Publicacion();
-        $publicacion->titulo=$request->titulo;
-        $publicacion->texto=$request->texto;
-        $publicacion->portada=1;
-        dd($publicacion);
-        $publicacion=Publicacion::create($publicacion);
-
-       // return new PublicacionResource($publicacion);
+        $publicacion = Publicacion::create($request->validated());
+        return new PublicacionResource($publicacion);
     }
 
     /**
      * Obtiene la publicacion especificada en su id
      */
-    public function show(Publicacion $publicacion)
+    public function show($id)
     {
+        $publicacion = Publicacion::with([
+            'imagenes',
+            'equipo',
+            'partido',
+            'patrocinador',
+            'jugador',
+            'reto',
+            'ong',
+            'pabellon'
+        ])->find($id);
+
+        if (!$publicacion) {
+            return response()->json(['error' => 'Publicación no encontrada'], 404);
+        }
+
         return new PublicacionResource($publicacion);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(StorePublicacionRequest $request, Publicacion $publicacion)
+    public function update(UpdatePublicacionRequest $request, $id)
     {
-        $datos = $request()->validated();
-        $publicacion->update($datos);
-        return new PublicacionResource($publicacion);
+       // Buscar el jugador por el ID
+       $publicacion = Publicacion::find($id);
+       // Verificar si el jugador existe
+       if (!$publicacion) {
+           return response()->json(['error' => 'Publicacion no encontrada'], 404);
+       }
+       // Validar los datos del request
+       $datos = $request->validated();
+       // Actualizar los datos del jugador
+       $publicacion->update($datos);
+       // Devolver una respuesta con el jugador actualizado
+       return response()->json(['message' => 'Publicacion actualizada correctamente', 'data' => $publicacion]);
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Publicacion $publicacion)
+    public function destroy($id)
     {
+        $publicacion = Publicacion::find($id);
+        if (!$publicacion) {
+            return response()->json(['error' => 'Publicacion no encontrada'], 404);
+        }
         $publicacion->delete();
         return response()->noContent();
     }
